Some weird errors when trying to write the interpreter
I am getting weird errors when writing the error handling for the
interpreter, as well as that I am using the wrong compiler, (32-bit
MinGW instead of 64). To switch compilers I just need to install some
libraries, for the lua problem, I am not entirely sure as to what is
causing it. lua_tostring(L, -1), where L represents the Lua virtual
machine and -1 represents the index on the stack.