Did a lot of things
I changed the coordinate system, which screwed up the drawing of a tri
colored triangle on the screen, as well as that I started to implement
audio, I also implemented the graphing, which basically calculates the
value of y at a given x with a given function. I also polished up some
interpreter stuff, added in getters to the window and a window resize
callback.