Fixed all the problems and made the graph look cool
The graph is looking cool, I discovered I was using the wrong type in
the for loop, and using a couple of sine waves I managed to make the
graph of a parabola look like a seagull. Aswell as color it in a very
cool way. This commit is written from my home computer, and I am using
the mingw64 compiler on it, and it is working somehow.