Was messing with how graphs are drawn and found scaling
So the variable in the graph named scale doesn't actually work, I found
it distorts the graph a fair bit. However, there is an OpenGL function
that allows for scaling to occur. It is called glScaled. If the number
is 1 then it is the regular scale, if the number is > 1 then it view
will be larger, if the number is < 1, the view will be smaller.