Made some features
I implemented things where the graph can be shifted over by a speed,
which can be incremented or decremented, as well as I added in scaling.
I also tried adding in some audio features, but the compiler just didn't
want to work with it. I also tried to add a feature where the color of a
graph can be specified by using a variable called c which is a table of
r, g, b. This would imply that files would need to be parsed in pairs of
lines so that if there are any line colors, then they do not get mixed
up. I also started working on the configuration file that can be used
customize the calculator.