The last commit did not go through
For some reason the last commit did not go through, but basically I
added in a keyboard listener so now I can change the scale and position
of the graph, as well as that I started working on a configuration file.
I also thought I would add in a variable that can be read by C++ that
will specify the color of the graph. Lastly I also started working on
the audio, but the compiler did not like it.