Some problems
Graphing of advanced functions have been done and I realized that I
cannot simply just put an x value, rather I need every y value as well.
For this I need to make an optimization where I get the gx and gy of the
camera, multiply with the scale value to get a range of x and y to
render from and yo.