Added in themes and started pondering comfort features
I added in themes and theme switching. I am also thinking of adding
comfort features, but I am also looking at the consequences if I do,
such as that I would need to make a parser, when the parser works for
x^2 and any other number, then I have to figure out how to make x be
able to become (x^2 + 2x + 1) and so on, or (x+1)(x-1).