Y axis time
I made it so that the calculator can now draw on both axis, this is
because I can check the function for either x or y and then change the
method of placing vertices. However, it cannot draw on both axis at the
same time, I don't have any ideas on how to do this as there is no easy
way to implement it. I also need to implement a way for the user to
define inequalities. For now I will just be writing some documentation
and cleaning up the code.