Both axis and inequalities
I added in support for equations on both axis, but it does take sometime
to process, so I added in a start zoom variable, which will allow for
the start zoom to be configured. I also added inequalities, they're a
bit rough around the edges, but they allow for the user to say for every
value of sin(x) that is > 0, and it will be applied. With = and != signs
a point should be placed. As well as for graphs that are on the y axis
they rely on the camera's x position rather then y. I plan to implement
sound and clean up the code a bit more.