Documented and changed graphing
I documented most of the code, just a simple definition of what a
function does and variables. I also changed the parsing to be done when
calculating the x and y of the point. I also added in a simple color
parser and added a few more entries to the point structure. I could not
test the code, but I will do this from my home computer.